Remove old beaconing code!
