+++ analysis/TBD/TBD.go
@@ -3,7 +3,7 @@ package TBD
 import (
 	"math"
 	"sort"
-	"strconv"
+	"sync"
 	"time"
 
 	"gopkg.in/mgo.v2/bson"
@@ -17,459 +17,340 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
-// Data coming from unique connection table aggregation
 type (
-	TBD struct {
-		db        string // database
-		resources *config.Resources
-		log       *log.Logger // system logger
+	empty struct{}
 
-		batch_size          int     // BatchSize
-		prefetch            float64 // Prefetch
-		mintime             int64   // minimum time
-		maxtime             int64   // maximum time
-		default_bucket_size float64 // default size of buckets
-		default_conn_thresh int     // default connections threshold
+	TBD struct {
+		db                string // database
+		resources         *config.Resources
+		defaultConnThresh int // default connections threshold
+		collectChannel    chan string
+		analysisChannel   chan *tbdAnalysisInput
+		writeChannel      chan *datatype_TBD.TBDAnalysisOutput
+		collectWg         sync.WaitGroup
+		analysisWg        sync.WaitGroup
+		writeWg           sync.WaitGroup
+		collectThreads    int
+		analysisThreads   int
+		writeThreads      int
+		log               *log.Logger // system Logger
+		minTime           int64       // minimum time
+		maxTime           int64       // maximum time
 	}
 
-	// Key provides a structure for keying the lookups
-	Key struct {
-		ID  bson.ObjectId
-		Src string
-		Dst string
+	tbdAnalysisInput struct {
+		src     string
+		dst     string
+		uconnID bson.ObjectId
+		ts      []int64
+		//dur []int64
+		//orig_bytes []int64
+		//resp_bytes []int64
 	}
 )
 
-// New produces a new TBD Module
+// Name gives the name of this module
 func New(c *config.Resources) *TBD {
 	return &TBD{
-		db:                  c.System.DB,
-		resources:           c,
-		log:                 c.Log,
-		batch_size:          c.System.BatchSize,
-		prefetch:            c.System.Prefetch,
-		default_bucket_size: c.System.TBDConfig.DefaultBucketSize,
-		default_conn_thresh: c.System.TBDConfig.DefaultConnectionThresh,
+		db:                c.System.DB,
+		resources:         c,
+		defaultConnThresh: c.System.TBDConfig.DefaultConnectionThresh,
+		log:               c.Log,
+		collectChannel:    make(chan string),
+		analysisChannel:   make(chan *tbdAnalysisInput),
+		writeChannel:      make(chan *datatype_TBD.TBDAnalysisOutput),
+		collectThreads:    2,
+		analysisThreads:   2,
+		writeThreads:      2,
 	}
 }
 
-// Write data out to mongodb
-func (t *TBD) writeData(
-	key Key,
-	rng int64,
-	size int64,
-	range_vals string,
-	fill float64,
-	spread float64,
-	sum int64,
-	score float64,
-	intervalkeys []int64,
-	intervalvals []int64,
-	maxkey int64,
-	maxval int64,
-	tss []int64) {
-
-	t.log.WithFields(log.Fields{
-		"uconn_id": key.ID,
-	}).Debug("Writing document")
-	session := t.resources.Session.Copy()
-	defer session.Close()
-
-	// Create a write object
-	obj := datatype_TBD.TBD{
-		Src:           key.Src,
-		Dst:           key.Dst,
-		UconnID:       key.ID,
-		Range:         rng,
-		Size:          size,
-		RangeVals:     range_vals,
-		Fill:          fill,
-		Spread:        spread,
-		Sum:           sum,
-		Score:         score,
-		Intervals:     intervalkeys,
-		InvervalCount: intervalvals,
-		TopInterval:   maxkey,
-		TopIntervalCt: maxval,
-		Tss:           tss,
-	}
-
-	err := session.DB(t.db).C(t.resources.System.TBDConfig.TBDTable).Insert(obj)
-	if err != nil {
-		t.log.WithFields(log.Fields{
-			"error": err.Error(),
-		}).Error("Error writing TBD results to database")
-	}
-	t.log.Debug("Document written")
-}
-
-func (t *TBD) createIntervalMaps(timestampList []int64) (map[int64]int64, map[int64]int64, []int64) {
-
-	// Get the end user selected bucket size
-	bucket_size := t.default_bucket_size
-
-	// Sort the timestamps in ascending order
-	var ts_list util.SortableInt64
-	ts_list = timestampList
-	sort.Sort(ts_list)
-
-	// Create a map that will act as a histogram of when connections occurred
-	time_buckets := make(map[int64]int64)
-
-	// Create a map that will act as a histogram for the amount of time between consecutive connections
-	interval_buckets := make(map[int64]int64)
-
-	// Create a list to hold sorted frequencies
-	var interval_list []int64
-
-	// Iterate over all the timestamps
-	for i := 0; i < len(ts_list)-1; i++ {
-
-		// Determine the number of seconds between two consecutive timestamps
-		interval := ts_list[i+1] - ts_list[i]
-
-		// Determine which time bucket the connection belongs in
-		time := int64(float64(ts_list[i]-ts_list[0]) / bucket_size)
-
-		// Counting the number of times timestamps show up in predetermined intervals of time
-		_, timeFound := time_buckets[time]
-		if timeFound {
-			time_buckets[time]++
-		} else {
-			time_buckets[time]++
-		}
-
-		// Count the number of times specific frequencies show up
-		if interval > 1 {
-			_, freqFound := interval_buckets[interval]
-			if freqFound {
-				interval_buckets[interval]++
-			} else {
-				interval_buckets[interval]++
-				interval_list = append(interval_list, interval)
-			}
-		}
-	}
-
-	return time_buckets, interval_buckets, interval_list
-}
-
-// TBDanalysis runs the TBD analysis on the dataset
-func (t *TBD) TBDanalysis(timestampMap map[Key][]int64) {
-	t.log.WithFields(log.Fields{
-		"count": len(timestampMap),
-	}).Debug("Entered analysis with count data points")
-
-	// Get the end user selected bucket size
-	bucket_size := t.default_bucket_size
-
-	// Get the threshold for the number of connections that need to be seen before
-	// a (src,dst,dpt) tuple is analyzed.
-	connection_thresh := t.default_conn_thresh
-	// Start result counter
-	// written := 0
-
-	var_map := make(map[Key]int64)
-	sum_map := make(map[Key]int64)
-	size_map := make(map[Key]int64)
-	range_map := make(map[Key][]int64)
-	fill_map := make(map[Key]int64)
-	spread_map := make(map[Key]int64)
-
-	// Iterate through the map
-	for key, ts_list_temp := range timestampMap {
-
-		// Sort the timestamps in ascending order
-		var ts_list util.SortableInt64
-		ts_list = ts_list_temp
-		sort.Sort(ts_list)
-
-		// Get the index for the last element in the list, since we use it in a few places
-		last := len(ts_list) - 1
-		first := 0
-
-		// Find the length of time between the last and first timestamp for this connection
-		sprd := ts_list[last] - ts_list[first]
-
-		time_buckets, interval_buckets, interval_list := t.createIntervalMaps(ts_list)
-
-		// Sort the keys (frequency intervals) for the frequency buckets in ascending order
-		sort.Sort(util.SortableInt64(interval_list))
-
-		// Number of frequencies. This will give us the number of frequency intervals
-		interval_length := len(interval_list)
-
-		// sum of values - This will give us the total number of connections
-		interval_sum := int64(0)
-		for i := 0; i < interval_length; i++ {
-			interval_sum += interval_buckets[int64(interval_list[i])]
-		}
-
-		// If there were more than connection_thresh total connections...
-		if len(ts_list) > connection_thresh && len(interval_list) > 0 {
-			// best (smallest) range of key values that make up X% of the
-			// freq_buckets histogram.
-			best := int64(31536000) // (one year in seconds)
-
-			best_sum := int64(0)
-
-			// The keys in freq_buckets that make up the best range
-			// for the variable above (best)
-			best_range := []int64{0}
-
-			// If there is only one frequency (ie every connection happens 2 seconds after the last)...
-			// Return that one frequency
-			if interval_length == 1 {
-				best = 1
-				best_range = []int64{interval_list[0]}
-				best_sum = interval_buckets[int64(interval_list[0])]
-			}
-
-			// We are trying to find the smallest range of values that has the largest
-			// "weight" in the histogram -- For example, trying to find the smallest
-			// range that makes up 85% of the total data.
-
-			// This loop iterates over the starting point for the range.
-			for i := 0; i < interval_length-1; i++ {
-
-				// Keep track of the sum of the number of connections having a freqeuncy
-				// in the current range.
-				mSum := interval_buckets[int64(interval_list[i])]
-
-				// Keep track of the range between high and low frequencies
-				mRange := int64(1)
-
-				// Check if the current range contains X% of the total data
-				if (float64(mSum)/float64(interval_sum) > 0.85) && (mRange < best) {
-					best = mRange
-					best_range = []int64{interval_list[i]}
-					best_sum = mSum
-				}
-
-				// Increment the end of the range
-				for j := i + 1; j < interval_length; j++ {
-					mSum += interval_buckets[int64(interval_list[j])]
-					mRange = interval_list[j] - interval_list[i]
-
-					// Check if this range contains X% of the total data
-					if (float64(mSum)/float64(interval_sum) > 0.85) && (mRange < best) {
-						best = mRange
-						best_range = []int64{interval_list[i], interval_list[j]}
-						best_sum = mSum
-					}
-				}
-			}
-			var_map[key] = best
-			size_map[key] = interval_sum //float64(len(ts_list))
-			range_map[key] = best_range
-			fill_map[key] = int64(len(time_buckets))
-			spread_map[key] = sprd
-			sum_map[key] = best_sum
-		}
-	}
-
-	t.log.Debug("Entering the presence check..")
-	for key, tsList := range timestampMap {
-		present := true
-
-		// Make sure all the necessary fields are available...
-		if _, ok := var_map[key]; !ok {
-			present = false
-		}
-		if _, ok := size_map[key]; !ok {
-			present = false
-		}
-		if _, ok := range_map[key]; !ok {
-			present = false
-		}
-		if _, ok := fill_map[key]; !ok {
-			present = false
-		}
-		if _, ok := sum_map[key]; !ok {
-			present = false
-		}
-
-		if present {
-			t.log.Debug("Preparing to write data")
-			// Extract all the necessary fields
-			rng := var_map[key]
-			size := size_map[key]
-			range_vals := strconv.Itoa(int(range_map[key][0]))
-			range_avg := float64(range_map[key][0])
-			if len(range_map[key]) > 1 {
-				range_min := strconv.Itoa(int(range_map[key][0]))
-				range_max := strconv.Itoa(int(range_map[key][1]))
-				range_vals = range_min + "--" + range_max
-				range_avg = (range_avg + float64(range_map[key][1])) / 2.0
-			}
-			fill := float64(fill_map[key]) / (float64(t.maxtime-t.mintime) / bucket_size)
-			spread := float64(spread_map[key]) / float64(t.maxtime-t.mintime)
-			sum := sum_map[key]
-
-			ideal_sum := float64(t.maxtime-t.mintime) / range_avg
-			alpha := float64(sum) / ideal_sum
-			if alpha > 1 {
-				alpha = 1
-			}
-			alpha = 1 - alpha
-
-			beta := float64(rng) / 600.0
-
-			if beta > 1 {
-				beta = 1
-			}
-
-			gamma := 1 - spread
-
-			score := math.Sqrt(alpha*alpha + beta*beta + gamma*gamma)
-
-			// Write out results with a score < 0
-			if score < 1.0 {
-
-				// Get interval lists for the interesting results
-				_, interval_buckets, _ := t.createIntervalMaps(tsList)
-				keys := make([]int64, 0, len(interval_buckets))
-				vals := make([]int64, 0, len(interval_buckets))
-
-				maxkey := int64(0)
-				maxval := int64(0)
-				for k, v := range interval_buckets {
-					keys = append(keys, k)
-					vals = append(vals, v)
-					if v > maxval {
-						maxval = v
-						maxkey = k
-					}
-				}
-
-				// Write the results.
-				t.log.Debug("Writing to the database")
-				t.writeData(key, rng, size, range_vals, fill*100, spread*100, sum, score, keys, vals, maxkey, maxval, tsList)
-			}
-
-		}
-
-	}
-
-}
-
-// Run runst the module against current configuration
+//Start the TBD process
 func (t *TBD) Run() {
-	t.log.Info("starting beacon hunt")
+	t.log.Info("Running beacon hunt")
 	session := t.resources.Session.Copy()
-	session.SetSocketTimeout(1 * time.Hour)
 	defer session.Close()
 
-	// Find First Time
-	t.log.Debug("Looking for first time")
+	//Find first time
+	t.log.Debug("Looking for first connection timestamp")
 	start := time.Now()
 
-	var d data.Conn
-	var first int64 = 100000000000
-	var last int64 = -1
+	var conn data.Conn
+	session.DB(t.db).
+		C(t.resources.System.StructureConfig.ConnTable).
+		Find(nil).Limit(1).Sort("ts").Iter().Next(&conn)
 
-	fliter := session.DB(t.db).C(t.resources.System.StructureConfig.ConnTable).
-		Find(nil).
-		Batch(t.batch_size).
-		Prefetch(t.prefetch).
-		Limit(1).
-		Sort("ts").
-		Iter()
+	t.minTime = conn.Ts
 
-	for fliter.Next(&d) {
-		first = d.Ts
-	}
+	t.log.Debug("Looking for last connection timestamp")
+	session.DB(t.db).
+		C(t.resources.System.StructureConfig.ConnTable).
+		Find(nil).Limit(1).Sort("-ts").Iter().Next(&conn)
 
-	// Find Last Time
-	t.log.Debug("Looking for last time")
-	fliter2 := session.DB(t.db).C(t.resources.System.StructureConfig.ConnTable).
-		Find(nil).
-		Batch(t.batch_size).
-		Prefetch(t.prefetch).
-		Limit(1).
-		Sort("-ts").
-		Iter()
+	t.maxTime = conn.Ts
 
-	for fliter2.Next(&d) {
-		last = d.Ts
-	}
-
-	t.mintime = first
-	t.maxtime = last
 	t.log.WithFields(log.Fields{
 		"time_elapsed": time.Since(start),
-		"first":        t.mintime,
-		"last":         t.maxtime,
-	}).Debug("First and last found")
+		"first":        t.minTime,
+		"last":         t.maxTime,
+	}).Debug("First and last timestamps found")
 
-	// Create list of all local addresses
-	var localAddresses []string
-	var current structure.Host
+	// add local addresses to channel
+	var host structure.Host
+	localIter := session.DB(t.db).
+		C(t.resources.System.StructureConfig.HostTable).
+		Find(bson.M{"local": true}).Iter()
 
-	localIter := session.DB(t.db).C(t.resources.System.StructureConfig.HostTable).
-		Find(bson.M{"local": true}).
-		Iter()
-
-	for localIter.Next(&current) {
-		localAddresses = append(localAddresses, current.Ip)
+	for i := 0; i < t.collectThreads; i++ {
+		t.collectWg.Add(1)
+		go t.collect()
 	}
 
-	t.log.WithFields(log.Fields{
-		"count": len(localAddresses),
-	}).Info("found all local addresses")
+	for i := 0; i < t.analysisThreads; i++ {
+		t.analysisWg.Add(1)
+		go t.analyze()
+	}
 
-	var keyList []Key
+	for i := 0; i < t.writeThreads; i++ {
+		t.writeWg.Add(1)
+		go t.write()
+	}
 
-	for _, ipAddress := range localAddresses {
-		var uc structure.UniqueConnection
+	for localIter.Next(&host) {
+		t.collectChannel <- host.Ip
+	}
+	t.log.Debug("Finding all source / destination pairs for analysis")
+	close(t.collectChannel)
+	t.collectWg.Wait()
+	t.log.Debug("Analyzing source / destination pairs")
+	close(t.analysisChannel)
+	t.analysisWg.Wait()
+	t.log.Debug("Finishing writing results to database")
+	close(t.writeChannel)
+	t.writeWg.Wait()
+}
+
+func (t *TBD) collect() {
+	session := t.resources.Session.Copy()
+	defer session.Close()
+	host, more := <-t.collectChannel
+	for more {
+		//grab all destinations related with this host
+		var uconn structure.UniqueConnection
 		destIter := session.DB(t.db).
 			C(t.resources.System.StructureConfig.UniqueConnTable).
-			Find(bson.M{"src": ipAddress}).Iter()
+			Find(bson.M{"src": host}).Iter()
 
-		for destIter.Next(&uc) {
-			if uc.ConnectionCount < t.default_conn_thresh {
+		for destIter.Next(&uconn) {
+			//skip the connection pair if they are under the threshold
+			if uconn.ConnectionCount < t.defaultConnThresh {
 				continue
 			}
 
-			newKey := Key{
-				ID:  uc.ID,
-				Src: uc.Src,
-				Dst: uc.Dst,
+			//create our new input
+			newInput := &tbdAnalysisInput{
+				uconnID: uconn.ID,
+				src:     uconn.Src,
+				dst:     uconn.Dst,
 			}
 
-			keyList = append(keyList, newKey)
+			//Grab connection data
+			var conn data.Conn
+			connIter := session.DB(t.db).
+				C(t.resources.System.StructureConfig.ConnTable).
+				Find(bson.M{"id_origin_h": uconn.Src, "id_resp_h": uconn.Dst}).
+				Iter()
+
+			for connIter.Next(&conn) {
+				newInput.ts = append(newInput.ts, conn.Ts)
+			}
+			t.analysisChannel <- newInput
+		}
+		host, more = <-t.collectChannel
+	}
+	t.collectWg.Done()
+}
+
+func (t *TBD) analyze() {
+	data, more := <-t.analysisChannel
+	for more {
+		sort.Sort(util.SortableInt64(data.ts))
+
+		//store the diff slice length since we use it a lot
+		length := len(data.ts) - 1
+
+		//find the duration of this connection
+		//perfect beacons should fill the observation period
+		duration := float64(data.ts[length]-data.ts[0]) /
+			float64(t.maxTime-t.minTime)
+
+		//find the delta times between the timestamps
+		var diff []int64 = make([]int64, length)
+		for i := 0; i < length; i++ {
+			diff[i] = data.ts[i+1] - data.ts[i]
 		}
 
+		//perfect beacons should have symmetric delta time distributions
+		//Bowley's measure of skew is used to check symmetry
+		sort.Sort(util.SortableInt64(diff))
+		bSkew := float64(0)
+		low := diff[toInt64(.25*float64(length-1))]
+		mid := diff[toInt64(.5*float64(length-1))]
+		high := diff[toInt64(.75*float64(length-1))]
+		bNum := low + high - 2*mid
+		bDen := high - low
+		if bDen != 0 {
+			bSkew = float64(bNum) / float64(bDen)
+		}
+
+		//perfect beacons should have very low dispersion around the
+		//median of their delta times
+		//Median Absolute Deviation About the Median
+		//is used to check dispersion
+		devs := make([]int64, length)
+		for i := 0; i < length; i++ {
+			devs[i] = abs(diff[i] - mid)
+		}
+
+		sort.Sort(util.SortableInt64(devs))
+		madm := devs[toInt64(.5*float64(length-1))]
+
+		//Store the range for human analysis
+		iRange := diff[length-1] - diff[0]
+
+		//get a list of the intervals found in the data,
+		//the number of times the interval was found,
+		//and the most occurring interval
+		intervals, intervalCounts, mode, modeCount := createCountMap(diff)
+
+		newOutput := &datatype_TBD.TBDAnalysisOutput{
+			UconnID:           data.uconnID,
+			TS_iSkew:          bSkew,
+			TS_iDispersion:    madm,
+			TS_duration:       duration,
+			TS_iRange:         iRange,
+			TS_iMode:          mode,
+			TS_iModeCount:     modeCount,
+			TS_intervals:      intervals,
+			TS_intervalCounts: intervalCounts,
+		}
+
+		//more skewed distributions recieve a lower score
+		//skew is mainly used as a tie breaker
+		alpha := 1.0 - math.Abs(bSkew)
+
+		//lower dispersion is better, cutoff dispersion scores at 30 seconds
+		beta := 1.0 - float64(madm)/30.0
+		if beta < 0 {
+			beta = 0
+		}
+		gamma := duration
+
+		//in order of ascending importance: skew, duration, dispersion
+		newOutput.TS_score = (alpha + beta + gamma) / 3.0
+
+		t.writeChannel <- newOutput
+		data, more = <-t.analysisChannel
+	}
+	t.analysisWg.Done()
+}
+
+func (t *TBD) write() {
+	session := t.resources.Session.Copy()
+	defer session.Close()
+
+	data, more := <-t.writeChannel
+	for more {
+		session.DB(t.db).C(t.resources.System.TBDConfig.TBDTable).Insert(data)
+		data, more = <-t.writeChannel
+	}
+	t.writeWg.Done()
+}
+
+//returns a distinct data array, data count array, the mode,
+//and the number of times the mode occured
+func createCountMap(data []int64) ([]int64, []int64, int64, int64) {
+	//create interval counts for human analysis
+	dataMap := make(map[int64]int64)
+	for _, d := range data {
+		dataMap[d]++
 	}
 
-	t.log.WithFields(log.Fields{
-		"count": len(keyList),
-	}).Info("found all connections above threshold")
+	distinct := make([]int64, len(dataMap))
+	counts := make([]int64, len(dataMap))
 
-	tsMap := make(map[Key][]int64)
-
-	longest := 0
-
-	for _, currKey := range keyList {
-		var curConn data.Conn
-		tsIter := session.DB(t.db).
-			C(t.resources.System.StructureConfig.ConnTable).
-			Find(bson.M{"id_origin_h": currKey.Src, "id_resp_h": currKey.Dst}).
-			Iter()
-
-		for tsIter.Next(&curConn) {
-			tsMap[currKey] = append(tsMap[currKey], curConn.Ts)
-		}
-
-		if longest < len(tsMap[currKey]) {
-			longest = len(tsMap[currKey])
-		}
+	i := 0
+	for k, v := range dataMap {
+		distinct[i] = k
+		counts[i] = v
+		i++
 	}
 
-	t.log.WithFields(log.Fields{
-		"most_connections": longest,
-	}).Info("map is built, preparing analysis")
-	t.TBDanalysis(tsMap)
+	mode := distinct[0]
+	max := counts[0]
+	for idx, count := range counts {
+		if count > max {
+			max = count
+			mode = distinct[idx]
+		}
+	}
+	return distinct, counts, mode, max
+}
 
-	return
+func GetViewPipeline(r *config.Resources, cuttoff float64) []bson.D {
+	return []bson.D{
+		{
+			{"$match", bson.D{
+				{"ts_score", bson.D{
+					{"$gt", cuttoff},
+				}},
+			}},
+		},
+		{
+			{"$lookup", bson.D{
+				{"from", r.System.StructureConfig.UniqueConnTable},
+				{"localField", "uconn_id"},
+				{"foreignField", "_id"},
+				{"as", "uconn"},
+			}},
+		},
+		{
+			{"$unwind", "$uconn"},
+		},
+		{
+			{"$sort", bson.D{
+				{"ts_score", -1},
+			}},
+		},
+		{
+			{"$project", bson.D{
+				{"ts_score", 1},
+				{"src", "$uconn.src"},
+				{"dst", "$uconn.dst"},
+				{"connection_count", "$uconn.connection_count"},
+				{"avg_bytes", "$uconn.avg_bytes"},
+				{"ts_iRange", 1},
+				{"ts_iMode", 1},
+				{"ts_iMode_count", 1},
+				{"ts_iSkew", 1},
+				{"ts_duration", 1},
+			}},
+		},
+	}
+}
+
+//TODO: Move these to util
+func abs(a int64) int64 {
+	if a >= 0 {
+		return a
+	}
+	return -a
+}
+
+func toInt64(f float64) int64 {
+	_, float := math.Modf(f)
+	if float > .5 {
+		return int64(math.Ceil(f))
+	}
+	return int64(f)
 }
