+++ commands/analyze.go
@@ -130,10 +130,6 @@ func analyze(inDb string, verboseFlag bool) {
 
 		d.BuildTBDCollection()
 
-		/////TESTING CODE
-		d.BuildTBDv2Collection()
-		/////END TESTING CODE
-
 		if verboseFlag {
 			fmt.Fprintf(os.Stdout,
 				"\t[-] Running blacklisted analysis\n")
