+++ commands/show-beacons.go
@@ -3,30 +3,29 @@ package commands
 import (
 	"fmt"
 	"os"
-	"sort"
-	"strings"
 	"text/template"
 
-	"gopkg.in/mgo.v2/bson"
-
 	"github.com/ocmdev/rita/config"
+	"github.com/ocmdev/rita/database"
+	"github.com/ocmdev/rita/datatypes/TBD"
 	"github.com/urfave/cli"
 )
 
-var beaconHeader = false
+var cutoffScore float64
 
 func init() {
 	command := cli.Command{
 		Name:  "show-beacons",
 		Usage: "print beacon information to standard out",
 		Flags: []cli.Flag{
-			cli.BoolFlag{
-				Name:        "no-header, r",
-				Usage:       "turn off the header row",
-				Destination: &beaconHeader,
-			},
 			humanFlag,
 			databaseFlag,
+			cli.Float64Flag{
+				Name:        "score, s",
+				Usage:       "change the beacon cutoff score",
+				Destination: &cutoffScore,
+				Value:       .7,
+			},
 		},
 		Action: showBeacons,
 	}
@@ -34,164 +33,67 @@ func init() {
 	bootstrapCommands(command)
 }
 
-type bcnresult struct {
-	Src           string        `bson:"src"`
-	Dst           string        `bson:"dst"`
-	UconnID       bson.ObjectId `bson:"uconn_id"`
-	Range         int64         `bson:"range"`
-	Size          int64         `bson:"size"`
-	RangeVals     string        `bson:"range_vals"`
-	Fill          float64       `bson:"fill"`
-	Spread        float64       `bson:"spread"`
-	Sum           int64         `bson:"range_size"`
-	Score         float64       `bson:"score"`
-	TopInterval   int64         `bson:"most_frequent_interval"`
-	TopIntervalCt int64         `bson:"most_frequent_interval_count"`
-	RangeMin      string
-	RangeMax      string
-}
-
-// The following is so that we can sort each of the beacons on thier score
-type bcnresultset []bcnresult
-
-func (slice bcnresultset) Len() int {
-	return len(slice)
-}
-
-func (slice bcnresultset) Less(i, j int) bool {
-	return slice[i].Score < slice[j].Score
-}
-
-func (slice bcnresultset) Swap(i, j int) {
-	slice[i], slice[j] = slice[j], slice[i]
-}
-
-// showBeacons shows all beacons for a given database
 func showBeacons(c *cli.Context) error {
-
 	if c.String("dataset") == "" {
 		return cli.NewExitError("No dataset was not specified", -1)
 	}
+	conf := config.InitConfig("")
+	conf.System.DB = c.String("dataset")
+
+	db := database.NewDB(conf)
+	data := db.GetTBDResultsView(cutoffScore)
 
 	if humanreadable {
-		return showBeaconsReport(c)
+		return showBeacon2Report(data)
 	}
 
-	return showBeaconCsv(c)
-
+	return showBeacon2Csv(data)
 }
 
-func showBeaconCsv(c *cli.Context) error {
-	tmpl := "{{.Score}},{{.Src}},{{.Dst}},{{.Range}},{{.Size}},{{.RangeMin}},{{.RangeMax}},"
-	tmpl += "{{.Fill}},{{.Spread}},{{.Sum}},{{.TopInterval}},{{.TopIntervalCt}}\n"
+func showBeacon2Report(data []TBD.TBDAnalysisView) error {
+	hdr := " Score |   Source IP    | Destination IP | Connections |"
+	hdr += " Avg. Bytes | Intvl Range | Top Intvl | Top Intvl Count |"
+	hdr += " Intvl Skew | Intvl Dispersion | Intvl Duration \n"
+	tmpl := `{{.TS_score | printf "%7.3f"}}` + " "
+	tmpl += `{{.Src | printf "%16s"}}` + " "
+	tmpl += `{{.Dst | printf "%16s"}}` + " "
+	tmpl += `{{.Connections | printf "%13d"}}` + " "
+	tmpl += `{{.AvgBytes | printf "%12.3f"}}` + " "
+	tmpl += `{{.TS_iRange | printf "%13d"}}` + " "
+	tmpl += `{{.TS_iMode | printf "%11d"}}` + " "
+	tmpl += `{{.TS_iModeCount | printf "%17d"}}` + " "
+	tmpl += `{{.TS_iSkew | printf "%12.3f"}}` + " "
+	tmpl += `{{.TS_iDispersion | printf "%18d"}}` + " "
+	tmpl += `{{.TS_duration | printf "%16.3f"}}` + "\n"
 
 	out, err := template.New("tbd").Parse(tmpl)
 	if err != nil {
-		panic(err)
+		return err
 	}
-
-	conf := config.InitConfig("")
-	if c.String("dataset") != "" {
-		conf.System.DB = c.String("dataset")
-	}
-
-	coll := conf.Session.DB(conf.System.DB).C(conf.System.TBDConfig.TBDTable)
-	iter := coll.Find(nil).Iter()
-
-	var res bcnresult
-	var allres bcnresultset
-
-	for iter.Next(&res) {
-		ranges := strings.Split(res.RangeVals, "--")
-		if len(ranges) == 1 {
-			res.RangeMin = ranges[0]
-			res.RangeMax = ranges[0]
-		} else {
-			res.RangeMin = ranges[0]
-			res.RangeMax = ranges[1]
-		}
-		allres = append(allres, res)
-	}
-
-	sort.Sort(allres)
-
-	for _, cres := range allres {
-		err := out.Execute(os.Stdout, cres)
+	fmt.Print(hdr)
+	for _, d := range data {
+		err := out.Execute(os.Stdout, d)
 		if err != nil {
 			fmt.Fprintf(os.Stdout, "ERROR: Template failure: %s\n", err.Error())
 		}
 	}
-
 	return nil
-
 }
 
-func showBeaconsReport(c *cli.Context) error {
-
-	if !c.Bool("no-header") {
-		hdr := "score\t          source\t            dest\trange\tsize\trange-min\trange-max\t  fill\tspread\t   sum\ttop-interval\ttop-interval-cnt\n"
-		hdr += "-----\t----------------\t----------------\t-----\t----\t---------\t---------\t------\t------\t------\t------------\t----------------\n"
-		fmt.Fprintf(os.Stdout, hdr)
-	}
-
-	tmpl := `{{.Score | printf "%.3f"}}`
-	tmpl += "\t{{.Src}}\t{{.Dst}}\t"
-	tmpl += `{{.Range | printf "%5d"}}`
-	tmpl += "\t" + `{{.Size | printf "%5d"}}` + "\t"
-	tmpl += "{{.RangeMin}}\t{{.RangeMax}}\t" + `{{.Fill | printf "%3.3f"}}`
-	tmpl += "\t" + `{{.Spread | printf "%3.3f"}}` + "\t" + `{{.Sum | printf "%6d"}}`
-	tmpl += "\t" + `{{.TopInterval | printf "%12d"}}` + "\t" + `{{.TopIntervalCt | printf "%16d"}}`
-	tmpl += "\n"
+func showBeacon2Csv(data []TBD.TBDAnalysisView) error {
+	tmpl := "{{.TS_score}},{{.Src}},{{.Dst}},{{.Connections}},{{.AvgBytes}},"
+	tmpl += "{{.TS_iRange}},{{.TS_iMode}},{{.TS_iModeCount}},"
+	tmpl += "{{.TS_iSkew}},{{.TS_iDispersion}},{{.TS_duration}}\n"
 
 	out, err := template.New("tbd").Parse(tmpl)
 	if err != nil {
-		panic(err)
+		return err
 	}
-
-	conf := config.InitConfig("")
-	if c.String("dataset") != "" {
-		conf.System.DB = c.String("dataset")
-	}
-
-	coll := conf.Session.DB(conf.System.DB).C(conf.System.TBDConfig.TBDTable)
-	iter := coll.Find(nil).Iter()
-
-	var res bcnresult
-	var allres bcnresultset
-
-	for iter.Next(&res) {
-		res = processBeaconbcnresult(res)
-		allres = append(allres, res)
-	}
-
-	sort.Sort(allres)
-
-	for _, cres := range allres {
-		err := out.Execute(os.Stdout, cres)
+	for _, d := range data {
+		err := out.Execute(os.Stdout, d)
 		if err != nil {
 			fmt.Fprintf(os.Stdout, "ERROR: Template failure: %s\n", err.Error())
 		}
 	}
-
 	return nil
-}
-
-func processBeaconbcnresult(r bcnresult) bcnresult {
-	r.Src = padAddr(r.Src)
-	r.Dst = padAddr(r.Dst)
-	mm := parseRangeVals(r.RangeVals)
-	r.RangeMin = mm[0]
-	r.RangeMax = mm[1]
-	return r
-}
-
-func parseRangeVals(rvals string) []string {
-	minmax := strings.Split(rvals, "--")
-	minmax[0] = leftPad(minmax[0], 8)
-	if len(minmax) == 1 {
-		minmax = append(minmax, minmax[0])
-		return minmax
-	}
-	minmax[1] = leftPad(minmax[1], 8)
-	return minmax
 }
